requirements.txt
uv
mcp
azure-communication-email
python-dotenv
